--- extensions/xj-search/raycast_search.js
@@ -1,176 +1,0 @@
-[
-  {
-    description: "",
-    icon: null,
-    title: "tce:bti_iocinfo:cn:prod",
-    url: "https://cloud.bytedance.net/tce/services/72116?cluster-type=cluster&module=cluster&page=1&page_size=10"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "tce:bti_iocinfo:i18n:prod",
-    url: "https://cloud-i18n.bytedance.net/tce/services/5731?cluster-type=cluster&module=cluster&page=1&page_size=10"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "tce:bti_iocinfo:boe:prod",
-    url: "https://cloud-boe.bytedance.net/tce/services/67619?cluster-type=cluster&module=cluster&page=1&page_size=10"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "bmq:iocshotstatus:cn:prod",
-    url: "https://cloud.bytedance.net/message-queue/topics/57794/detail?region=online&active_tab=consumer"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "bmq:ti_event:cn:prod",
-    url: "https://cloud.bytedance.net/message-queue/topics/57438/detail?region=online&active_tab=consumer"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "faas:ti_event_to_bytees:cn:prod",
-    url: "https://cloud.bytedance.net/faas/function/4sybdfmq/cluster/detail?cluster=faas-cn-north&region=cn-north"
-  },
-  {
-    description: "security.ioc.threatbk_lookup",
-    icon: null,
-    title: "faas:ioc_threatbk_lookup:cn:prod",
-    url: "https://cloud.bytedance.net/faas/function/ohc6zmth/detail"
-  },
-  {
-    description: "security.ioc.threatbk_lookup",
-    icon: null,
-    title: "git_repo:ioc_threatbk_lookup",
-    url: "https://code.byted.org/security/ioc_threatbk_lookup"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "bmq:security_ti_threatb_event:cn:prod",
-    url: "https://cloud.bytedance.net/message-queue/topics/993519/detail?region=online"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "bmq:ti_es:cn:prod",
-    url: "https://cloud.bytedance.net/message-queue/topics/931447/detail?region=online"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "bmq:bti_eval_engine_trigger:cn:prod",
-    url: "https://cloud.bytedance.net/message-queue/topics/57434/detail?region=online&active_tab=consumer"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "event:ioc_event:cn:prod",
-    url: "https://cloud.bytedance.net/eventbus_cn/eventv2/ioc_event?region=CN"
-  },
-  {
-    description: "ti-sec.bytedance.net",
-    icon: null,
-    title: "netlink:ti_sec:cn:prod",
-    url: "https://cloud.bytedance.net/netlink/main/namespace/insert-manage?ns=7452&show_type=domain&tdm_domain=ti-sec.bytedance.net&tdm_servername=26208"
-  },
-  {
-    description: "ioc.bytedance.net",
-    icon: null,
-    title: "netlink:ioc:cn:prod",
-    url: "https://cloud.bytedance.net/netlink/main/namespace/insert-manage?ns=4594"
-  },
-  {
-    description: "security.bti.iocinforestful",
-    icon: null,
-    title: "tce:iocinforestful:cn:prod",
-    url: "https://cloud.bytedance.net/tce/services/73781"
-  },
-  {
-    description: "security.bti.iocinforestful",
-    icon: null,
-    title: "git_repo:iocinforestful",
-    url: "https://code.byted.org/security/iocinforestful"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "faas:analysis_engine:cn:prod",
-    url: "https://cloud.bytedance.net/faas/function/it1kw8xs/cluster/detail?cluster=faas-cn-north&region=cn-north"
-  },
-  {
-    description: "security.bti.eval_engine",
-    icon: null,
-    title: "faas:bti_eval_engine:cn:prod",
-    url: "https://cloud.bytedance.net/faas/function/iwhd0a2e/cluster/detail"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "dorado:space:cn:prod",
-    url: "https://data.bytedance.net/dorado/development/node/101716761?project=cn_1233&version=-1"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "dorado:ioc_api_log_offline:cn:prod",
-    url: "https://data.bytedance.net/dorado/development/node/104309708?project=cn_1233"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "hive:ti_ioc_log_offline:cn:prod",
-    url: "https://data.bytedance.net/coral/datamap/detail?groupName=default&guid=38e36edc-c2d4-4ee2-a08f-adca2b7a50e6&qualifiedName=HiveTable%3A%2F%2F%2Fsecurity%2Fti_ioc_log_offline%400&subTab=schema#group=default"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "dorado:ods_bti_iocshots:cn:prod",
-    url: "https://data.bytedance.net/dorado/development/node/111335843?project=cn_1233"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "hive:ods_bti_iocshots:cn:prod",
-    url: "https://data.bytedance.net/coral/datamap/detail/table_info/HiveTable/security_threat_intelligence/ods_bti_iocshots/schema#group=default"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "clickhouse:bti_iocshots:cn:prod",
-    url: "https://data.bytedance.net/coral/datamap/detail?groupName=default&qualifiedName=ClickhouseTable%3A%2F%2F%2Fbytesecurity%2Fbti_iocshots%400&subTab=schema&tab=table_info#group=default"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "clickhouse:security_ti_event:cn:prod",
-    url: "https://data.bytedance.net/coral/datamap/detail?groupName=default&qualifiedName=ClickhouseTable%3A%2F%2F%2Fbytesecurity%2Fsecurity_ti_event%400&subTab=schema&tab=table_info#group=default"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "databus:security_ioc_log:cn:prod",
-    url: "https://cloud.bytedance.net/databus/cluster/cluster_detail?cluster=security_ioc_log&region=cn-north-1"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "bmq:security_bti_update:cn:prod",
-    url: "https://cloud.bytedance.net/message-queue/topics/59500/detail?region=online&active_tab=consumer"
-  },
-  {
-    description: null,
-    icon: null,
-    title: "tce:bti_platform:cn:prod",
-    url: "https://cloud.bytedance.net/tce/services/81769"
-  },
-  {
-    description: "bytedance.bytedoc.ti",
-    icon: null,
-    title: "mongo:ti:cn:prod",
-    url: "https://cloud.bytedance.net/bytedoc/classic/database/ti/overview?region=CN"
-  }
-]
