+++ extensions/xj-search/src/index.tsx
@@ -1,10 +1,10 @@
 import { ActionPanel, List, Action, LocalStorage, getPreferenceValues } from "@raycast/api";
-import * as yaml from "js-yaml";
-import { readdirSync, readFileSync } from "fs";
+import { UrlInfo } from "./interface";
 import { useState, useEffect } from "react";
 import path from "path";
 import { Fzf } from "fzf";
 import fs from "fs";
+import { convertYamlToJson } from "./utils";
 
 const raycastKey = "raycast_search";
 
@@ -30,86 +30,6 @@ function fzfSearch(data: UrlInfo[], query: string): UrlInfo[] {
   });
 
   return fzf.find(query).map((result) => result.item);
-}
-
-// Types and interfaces
-interface Task {
-  vars: { [key: string]: string };
-  desc: string;
-  cmd: string;
-}
-function readAndParseYaml(filePath: string): YamlFile | undefined {
-  try {
-    const yamlContent = readFileSync(filePath, "utf8");
-    return yaml.load(yamlContent) as YamlFile;
-  } catch (error) {
-    console.warn(`Error reading YAML file: ${filePath}`, error);
-    return undefined;
-  }
-}
-
-interface YamlFile {
-  vars: { namespace: string; icon?: string };
-  tasks: { [key: string]: Task };
-}
-
-interface UrlInfo {
-  title: string;
-  description?: string;
-  icon: string;
-  url: string;
-}
-const END_WITH = "-chrome-taskfile.yaml";
-
-// Helper Functions
-function getYamlFiles(folderPath: string): string[] {
-  return readdirSync(folderPath).filter((file) => file.endsWith(END_WITH));
-}
-
-function convertToUrlData(namespace2task: { [key: string]: { [key: string]: Task } }): unknown[] {
-  return Object.entries(namespace2task).map(([namespace, tasks]) => ({
-    namespace,
-    urlsData: Object.entries(tasks).map(([taskName, task]) => ({
-      title: taskName,
-      description: task.desc || null,
-      icon: task.vars?.icon || "🔍",
-      url: task.vars?.URL,
-    })),
-  }));
-}
-
-function processYamlFile(
-  folderPath: string,
-  fileName: string,
-  accumulator: { [key: string]: { [key: string]: Task } },
-) {
-  const filePath = path.join(folderPath, fileName);
-  const parsedYaml = readAndParseYaml(filePath);
-
-  if (!parsedYaml || !parsedYaml.vars || !parsedYaml.tasks) {
-    console.warn(`Skipping invalid YAML file: ${fileName}`);
-    return;
-  }
-
-  const namespace = parsedYaml.vars.namespace || fileName.replace(END_WITH, "");
-  accumulator[namespace] = { ...accumulator[namespace], ...parsedYaml.tasks };
-}
-
-function convertYamlToJson(folderPath: string): string {
-  console.log("Converting YAML to JSON");
-
-  try {
-    const yamlFiles = getYamlFiles(folderPath);
-    console.log("Processing YAML files:", yamlFiles);
-
-    const namespace2task: { [key: string]: { [key: string]: Task } } = {};
-    yamlFiles.forEach((fileName) => processYamlFile(folderPath, fileName, namespace2task));
-
-    return JSON.stringify(convertToUrlData(namespace2task));
-  } catch (error) {
-    console.error("Error processing YAML files:", error);
-    return JSON.stringify([]);
-  }
 }
 
 function useRaycastData() {
