--- extensions/xj-search/test.json
@@ -1,9 +1,0 @@
-[
-    {
-        "description": null,
-        "icon": null,
-        "title": "TI_ES_CN_PROD_URL",
-        "url": "https://cloud.bytedance.net/tce/services/72116?cluster-type=cluster&module=cluster&page=1&page_size=10",
-        "version": "3"
-    }
-]
